expose xml decoder to the world
